Fixes for fs/fstat tests on NuttX/TizenRT.  - advanced libtuv (open(directory) provided)  - advanced tizenrt (fstat and statfs added)  - broken test/run_pass/test_fs_fstat.js into ../test_fs_fstat_{dir,file}.js
Tested with said tests on TizenRT.

IoT.js-DCO-1.0-Signed-off-by: Tomasz Wozniak t.wozniak@samsung.com
