+++ test/testsets.json
@@ -23,8 +23,9 @@
     { "name": "test_events_uncaught_error.js", "uncaught": true },
     { "name": "test_fs_exists.js" },
     { "name": "test_fs_exists_sync.js" },
-    { "name": "test_fs_fstat.js", "skip": ["nuttx", "tizenrt"], "reason": "not implemented for nuttx/TizenRT" },
-    { "name": "test_fs_fstat_sync.js", "skip": ["nuttx", "tizenrt"], "reason": "not implemented for nuttx/TizenRT" },
+    { "name": "test_fs_fstat_dir.js" },
+    { "name": "test_fs_fstat_file.js" },
+    { "name": "test_fs_fstat_sync.js" },
     { "name": "test_fs_mkdir_rmdir.js", "skip": ["linux", "nuttx"], "reason": "[linux]: flaky on Travis, [nuttx]: implemented, run manually in default configuration" },
     { "name": "test_fs_open_close.js", "skip": ["nuttx"], "reason": "not implemented for nuttx" },
     { "name": "test_fs_readdir.js" },
